refactor: update truncateAddress imports and refactor utils index/imports
