+++ src/features/domain-preview/Options/Options.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
+import { getDomainId } from '../../../lib/util';
 import { useWeb3 } from '../../../lib/hooks/useWeb3';
-import { getDomainId } from '../../../lib/util/domains/domains';
 import { useDomainData } from '../../../lib/hooks/useDomainData';
 
 import { MoreNFTOptions } from '../../ui/MoreNFTOptions';
