+++ src/features/view-bids/ViewBids.tsx
@@ -2,8 +2,7 @@ import { FC, useState } from 'react';
 
 import { useWeb3 } from '../../lib/hooks/useWeb3';
 import { useViewBidsData } from './useViewBidsData';
-import { sortBidsByTime } from '../../lib/util/bids/bids';
-import { formatEthers } from '../../lib/util';
+import { sortBidsByTime, formatEthers } from '../../lib/util';
 import { truncateAddress } from '@zero-tech/zapp-utils/formatting/addresses';
 import { Bid } from '@zero-tech/zauction-sdk';
 import moment from 'moment';
