+++ src/features/view-nft/Actions/Actions.tsx
@@ -2,9 +2,7 @@ import { useWeb3 } from '../../../lib/hooks/useWeb3';
 import { useBuyNowPrice } from '../../../lib/hooks/useBuyNowPrice';
 import { useBidData } from '../../../lib/hooks/useBidData';
 import { useDomainMetrics } from '../../../lib/hooks/useDomainMetrics';
-import { formatEthers } from '../../../lib/util/number/number';
-import { getUserBids } from '../../../lib/util/bids/bids';
-import { getDomainId } from '../../../lib/util/domains/domains';
+import { formatEthers, getUserBids, getDomainId } from '../../../lib/util';
 import { useDomainMetadata } from '../../../lib/hooks/useDomainMetadata';
 import { useDomainData } from '../../../lib/hooks/useDomainData';
 import { usePaymentToken } from '../../../lib/hooks/usePaymentToken';
