+++ src/lib/util/index.ts
@@ -1,2 +1,4 @@
+export * from './bids';
+export * from './ipfs';
 export * from './number';
 export * from './domains';
